LayoutTests/platform/chromium/fast/events/rtl-scrollbar.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../../fast/js/resources/js-test-pre.js"></script>
</head>
<body style="margin:0">
<div id="overflow" dir="rtl" style="border:2px solid black;overflow:auto;height:400px;width:400px; position:absolute;">
<div style="background-color:red;height:720px"></div>
<div style="background-color:green;height:1600px"></div>
</div>

<script>
description('Test that we can scroll down an RTL element with its left-side scrollbar.');

var scrollTop = document.getElementById('overflow').scrollTop;

if (window.testRunner)
    testRunner.waitUntilDone();

if (window.eventSender) {
    var node = document.getElementById('overflow');
    eventSender.mouseMoveTo(node.offsetLeft + 5, node.offsetTop + node.offsetHeight - 50);
    eventSender.mouseDown();
    eventSender.mouseUp();
    setTimeout(finished, 0);
}

function finished()
{
    shouldBeTrue('document.getElementById(\'overflow\').scrollTop > scrollTop');
    window.testRunner.notifyDone();
}

var successfullyParsed = true;
</script>
<script src="../../../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
